+++ demo/main.ts
@@ -2,7 +2,7 @@ import tmpl from "../mod.ts";
 
 const env = tmpl();
 
-const template = await env.load("./main.tmpl");
+const template = await env.load("./one.vto");
 
 const result = await template({
   title: "This is an example",