+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [Unreleased]
+### Fixed
+- Template literals in includes [#1].
+- Chain errors using the `cause` property.
+
 ## [0.3.1] - 2023-06-14
 ### Fixed
 - Improved errors.
@@ -27,6 +32,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ## [0.1.0] - 2023-06-04
 First version
 
+[#1]: https://github.com/oscarotero/vento/issues/1
+
+[Unreleased]: https://github.com/oscarotero/vento/compare/v0.3.1...HEAD
 [0.3.1]: https://github.com/oscarotero/vento/compare/v0.3.0...v0.3.1
 [0.3.0]: https://github.com/oscarotero/vento/compare/v0.2.0...v0.3.0
 [0.2.0]: https://github.com/oscarotero/vento/compare/v0.1.0...v0.2.0